requirements.txt
ipywidgets
ftfy
jupyter
jupyterlab

numpy
pillow
opencv-python

torch
torchvision
diffusers
transformers